SCHEDULING-1748 Support running stress tests on localhost without using ssh
Problem:
Currently stress tests framework always uses ssh command to start all
processes, even if command is executed on localhost. It would be good to
don't use ssh if it isn't really needed.

Fix:
Several places where modified to make it work on localhost without ssh:
- changed 'SSHProcessExecutor.createExecutor' to don't use ssh if target
host is localhost. After this method
'DeploymentTestUtils.isPathIsAvailable' started to fail for paths on
localhost. This method tries to execute 'cd' command with given path,
but 'cd' is a built-in shell command, so it doesn't work when it is run
outside of the shell. To make it work for both cases I changed method
'DeploymentTestUtils.isPathIsAvailable' to use 'test' command instead of
'cd'.
- changed method HostTestEnv.createFileInEnv to use 'cp' instead of
'rsync' when target file is on the localhost
- changed TestRMDeployer to deploy nodes using LocalInfrastructure if
all nodes are deployed on localhost
- changed TestDeployer to don't kill processes starting RM and Scheduler
in case if they are run on localhost

+++ performance/src/org/ow2/proactive/tests/performance/deployment/TestDeployer.java
@@ -145,7 +145,9 @@ public abstract class TestDeployer {
             error = t;
             return null;
         } finally {
-            executor.killProcess();
+            if (executor.isRunningRemotely()) {
+                executor.killProcess();
+            }
 
             if (error != null) {
                 System.out.println("Error during server deployment: " + error);
